requirements.txt
reportlab>=4.0.0
python-docx>=1.1.0
PyQt6>=6.6.0
pandas>=2.0.0
openpyxl>=3.1.0
Pillow>=10.0.0
pdfplumber>=0.10.0
PyPDF2>=3.0.0
